Move route handlers into enamel.api package
This is a first step in having the api live in its own package.

To cover the /servers route across this change a servers.yaml
test file was added.
